Added current developmemnt tag version file
- Created file to remember in which tag version of the project
development is taking palce right now. OR which version is found
at a specific commit or on a branch, if the same version was used
throughout the entire branch. My memory is not good so I won't be
able to remember the verison until it is done and I can tag it.